refactor(clawdstrike-policy-event): extract construct/ from edr/receipt/mod.rs
